refactor payload constants
