+++ pkg/api/networkservice/payload/constants.go
@@ -1,5 +1,7 @@
 // Copyright (c) 2020 Cisco and/or its affiliates.
 //
+// Copyright (c) 2020 Doc.ai and/or its affiliates.
+//
 // SPDX-License-Identifier: Apache-2.0
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,10 +16,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package ethernet provides constants and helpers for the Ethernet payload
-package ethernet
+// Package payload provides constants for NetworkService.Payload
+package payload
 
 const (
-	// PAYLOAD - value for an Ethernet payload
-	PAYLOAD = "ethernet"
+	// Ethernet is ethernet payload value for NetworkService.Payload
+	Ethernet = "ETHERNET"
+	// IP is IP payload value for NetworkService.Payload
+	IP = "IP"
 )
